Moving cache service to services folder
